+++ imaging/imageRendering.ts
@@ -55,14 +55,14 @@ import { setPixelShift } from "./loaders/dsaImageLoader";
  * If seriesId is passed as argument only imageIds of the series are purged from internal cache
  * @instance
  * @function clearImageCache
- * @param {String} seriesId - The id of the serie
+ * @param {String} uniqueUID - The uniqueUID of the series
  */
-export const clearImageCache = function (seriesId?: string) {
+export const clearImageCache = function (uniqueUID?: string) {
   const t0 = performance.now();
-  if (seriesId) {
+  if (uniqueUID) {
     let series = store.get("series");
-    if (has(series, seriesId)) {
-      each(series[seriesId].imageIds, function (imageId: string) {
+    if (has(series, uniqueUID)) {
+      each(series[uniqueUID].imageIds, function (imageId: string) {
         if (cornerstone.imageCache.cachedImages.length > 0) {
           try {
             cornerstone.imageCache.removeImageLoadObject(imageId);
@@ -76,8 +76,8 @@ export const clearImageCache = function (seriesId?: string) {
         }
       });
 
-      store.removeSeriesId(seriesId);
-      logger.info("Uncached images for ", seriesId);
+      store.removeImageIds(uniqueUID);
+      logger.info("Uncached images for ", uniqueUID);
     }
   } else {
     cornerstone.imageCache.purgeCache();
@@ -100,7 +100,7 @@ export function loadAndCacheImage(
 ): Promise<true> {
   const t0 = performance.now();
   // add serie's imageIds into store
-  store.addSeriesId(series.seriesUID, series.imageIds);
+  store.addImageIds(series.uniqueUID, series.imageIds);
   const imageId: string | undefined = series.imageIds[imageIndex];
 
   const cachePromise = new Promise<true>((resolve, reject) => {
@@ -111,7 +111,7 @@ export function loadAndCacheImage(
         const t1 = performance.now();
         logger.debug(`Call to cacheImages took ${t1 - t0} milliseconds.`);
         logger.debug(
-          `Cached image with index ${imageIndex} for ${series.seriesUID}`
+          `Cached image with index ${imageIndex} for ${series.uniqueUID}`
         );
         resolve(true);
       });
@@ -143,15 +143,15 @@ export function loadAndCacheImages(
   const t0 = performance.now();
   let cachingCounter = 0;
   const response = {
-    seriesId: series.seriesUID,
+    seriesId: series.uniqueUID,
     loading: 0,
     series: {} as Series
   };
   callback(response);
   // add serie's imageIds into store
-  store.addSeriesId(series.seriesUID, series.imageIds);
+  store.addImageIds(series.uniqueUID, series.imageIds);
   // add serie's caching progress into store
-  setStore(["progress", series.seriesUID, 0]);
+  setStore(["progress", series.uniqueUID, 0]);
 
   function updateProgress() {
     cachingCounter += 1;
@@ -159,11 +159,11 @@ export function loadAndCacheImages(
       (cachingCounter / series.imageIds.length) * 100
     );
     response.loading = cachingPercentage;
-    setStore(["progress", series.seriesUID, cachingPercentage]);
+    setStore(["progress", series.uniqueUID, cachingPercentage]);
     if (cachingCounter == series.imageIds.length) {
       const t1 = performance.now();
       logger.debug(`Call to cacheImages took ${t1 - t0} milliseconds.`);
-      logger.debug(`Cached images for ${series.seriesUID}`);
+      logger.debug(`Cached images for ${series.uniqueUID}`);
       response.series = series;
     }
   }
@@ -226,19 +226,19 @@ export const renderDICOMPDF = function (
   }
 
   // this by default is the uniqueId of the rendered stack
-  const storedSeriesUID = store.get(["viewports", id, "seriesUID"]);
-  const seriesUID = seriesStack.uniqueUID || seriesStack.seriesUID;
-  const isSeriesUIDChanged = storedSeriesUID !== seriesUID;
-  if (isSeriesUIDChanged) {
-    logger.debug(`SeriesUID changed from ${storedSeriesUID} to ${seriesUID}`);
+  const storedUniqueUID = store.get(["viewports", id, "uniqueUID"]);
+  const uniqueUID = seriesStack.uniqueUID || seriesStack.seriesUID;
+  const isUniqueUIDChanged = storedUniqueUID !== uniqueUID;
+  if (isUniqueUIDChanged) {
+    logger.debug(`UniqueUID changed from ${storedUniqueUID} to ${uniqueUID}`);
   }
 
   let renderPromise = new Promise<true>(async (resolve, reject) => {
     let image: Instance | null = seriesStack.instances[seriesStack.imageIds[0]];
     const SOPUID = image.dataSet?.string("x00080016");
 
-    if (isSeriesUIDChanged) {
-      setStore(["seriesUID", element.id, seriesUID]);
+    if (isUniqueUIDChanged) {
+      setStore(["uniqueUID", element.id, uniqueUID]);
     }
 
     // check sopUID in order to detect pdf report array
@@ -300,7 +300,7 @@ export const renderDICOMPDF = function (
           if (element) {
             const fileManager = getFileManager();
             if (fileManager) {
-              store.addSeriesId(seriesUID, Object.values(fileManager));
+              store.addImageIds(uniqueUID, Object.values(fileManager));
             } else {
               logger.error("FileManager is null");
             }
@@ -432,7 +432,7 @@ export const renderWebImage = function (
 /**
  * Unrender an image on a html div using cornerstone
  * Remove mouse listeners
- * Remove seriesUID from viewport store
+ * Remove uniqueUID from viewport store
  * Remove ready flag from viewport store
  * @instance
  * @function disableViewport
@@ -449,33 +449,31 @@ export const disableViewport = function (elementId: string | HTMLElement) {
   toggleMouseToolsListeners(element, true);
   cornerstone.disable(element);
   const id: string = isElement(elementId) ? element.id : (elementId as string);
-  setStore(["seriesUID", id, undefined]); // remove seriesUID from viewport store
+  setStore(["uniqueUID", id, undefined]); // remove uniqueUID from viewport store
   setStore(["ready", id, false]); // set ready to false in viewport store
 };
 
 /**
  * Unrender an image on a html div using cornerstone
  * Remove mouse listeners
- * Remove seriesUID from viewport store
+ * Remove uniqueUID from viewport store
  * Remove ready flag from viewport store
  * Remove image from cornerstone cache
  * Delete viewport from store
  * @instance
  * @function unloadViewport
  * @param {String} elementId - The html div id used for rendering or its DOM HTMLElement
- * @param {String} seriesId - The id of the serie
  */
-export const unloadViewport = function (elementId: string, seriesId: string) {
-  disableViewport(elementId);
-
-  if (!seriesId) {
-    logger.warn(
-      "seriesId not provided, use disableViewport if you do not want to uncache images"
-    );
+export const unloadViewport = function (elementId: string) {
+  const uniqueUID = store.get(["viewports", elementId, "uniqueUID"]);
+  if (!uniqueUID) {
+    logger.error("no uniqueUID found in store");
+    return;
   }
+  disableViewport(elementId);
   // remove images from cornerstone cache
-  if (seriesId && has(store.get("series"), seriesId)) {
-    clearImageCache(seriesId);
+  if (has(store.get("series"), uniqueUID)) {
+    clearImageCache(uniqueUID);
   }
   store.deleteViewport(elementId);
 };
@@ -550,11 +548,11 @@ export const renderImage = function (
   }
 
   // this by default is the uniqueId of the rendered stack
-  const storedSeriesUID = store.get(["viewports", id, "seriesUID"]);
-  const isSeriesUIDChanged = storedSeriesUID !== data.seriesUID;
-  if (isSeriesUIDChanged) {
+  const storedUniqueUID = store.get(["viewports", id, "uniqueUID"]);
+  const isUniqueUIDChanged = storedUniqueUID !== data.uniqueUID;
+  if (isUniqueUIDChanged) {
     logger.debug(
-      `SeriesUID changed from ${storedSeriesUID} to ${data.seriesUID}`
+      `uniqueUID changed from ${storedUniqueUID} to ${data.uniqueUID}`
     );
   }
 
@@ -669,9 +667,9 @@ export const renderImage = function (
           );
         }
 
-        // if seriesUID has changed update the value into the store
-        if (isSeriesUIDChanged) {
-          setStore(["seriesUID", element.id, data.seriesUID]);
+        // if uniqueUID has changed update the value into the store
+        if (isUniqueUIDChanged) {
+          setStore(["uniqueUID", element.id, data.uniqueUID]);
           if (renderOptions.scale === undefined) {
             viewport.scale = data.default?.scale || viewport.scale;
             logger.debug(
@@ -695,7 +693,7 @@ export const renderImage = function (
               viewport.rotation
             );
           }
-          // if the seriesUID has changed, update the viewport voi values
+          // if the uniqueUID has changed, update the viewport voi values
           // with the default values from the series
           // if the voi is not defined in the renderOptions
           if (renderOptions.voi === undefined) {
@@ -1344,7 +1342,7 @@ const getSeriesData = function (
   };
   type SeriesData = StoreViewport;
   const data: RecursivePartial<SeriesData> = {};
-  data.seriesUID = series.uniqueUID || series.seriesUID; //case of resliced series
+  data.uniqueUID = series.uniqueUID || series.seriesUID; //case of resliced series
   data.modality = series.modality;
 
   if (series.isMultiframe) {