+++ imaging/tools/default.ts
@@ -100,7 +100,7 @@ const DEFAULT_TOOLS: {
     cleanable: false,
     defaultActive: true,
     class: "WwwcTool",
-    // sync: "wwwcSynchronizer",
+    sync: "wwwcSynchronizer",
     description: "Change image contrast",
     shortcut: "ctrl-m",
     type: "utils"
@@ -116,7 +116,7 @@ const DEFAULT_TOOLS: {
     cleanable: false,
     defaultActive: false,
     class: "WwwcRegionTool",
-    // sync: "wwwcSynchronizer",
+    sync: "wwwcSynchronizer",
     description: "Change image contrast based on selected region",
     shortcut: "ctrl-m",
     type: "utils"
@@ -130,7 +130,7 @@ const DEFAULT_TOOLS: {
     },
     cleanable: false,
     defaultActive: false,
-    // sync: "wwwcSynchronizer",
+    sync: "wwwcSynchronizer",
     type: "utils",
     name: "WwwcRemoveRegion",
     class: "WwwcRemoveRegionTool",